Moves EDA Project and Inventories inside a Resources directory meet ansible-ui directory convention
