+++ frontend/eda/EventDrivenRouter.tsx
@@ -5,12 +5,12 @@ import Dashboard from './dashboard/Dashboard';
 import { EditExecutionEnvironment } from './execution-environments/EditExecutionEnvironment';
 import { ExecutionEnvironmentDetails } from './execution-environments/ExecutionEnvironmentDetails';
 import { ExecutionEnvironments } from './execution-environments/ExecutionEnvironments';
-import { EditInventory } from './inventories/EditInventory';
-import { Inventories } from './inventories/Inventories';
-import { InventoryDetails } from './inventories/InventoryDetails';
-import { EditProject } from './projects/EditProject';
-import { ProjectDetails } from './projects/ProjectDetails';
-import { Projects } from './projects/Projects';
+import { EditInventory } from './Resources/inventories/EditInventory';
+import { Inventories } from './Resources/inventories/Inventories';
+import { InventoryDetails } from './Resources/inventories/InventoryDetails';
+import { EditProject } from './Resources/projects/EditProject';
+import { ProjectDetails } from './Resources/projects/ProjectDetails';
+import { Projects } from './Resources/projects/Projects';
 import { EditRulebookActivation } from './rulebook-activations/EditRulebookActivation';
 import { RulebookActivationDetails } from './rulebook-activations/RulebookActivationDetails';
 import { RulebookActivations } from './rulebook-activations/RulebookActivations';
