+++ frontend/eda/Resources/projects/ProjectDetails.tsx
@@ -1,13 +1,13 @@
 import { DropdownPosition, PageSection } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { PageActions, PageHeader, PageLayout } from '../../../framework';
-import { Scrollable } from '../../../framework/components/Scrollable';
-import { TableDetails } from '../../../framework/PageTable/PageTableDetails';
-import { useSettings } from '../../../framework/Settings';
-import { useGet } from '../../common/useItem';
-import { RouteE } from '../../Routes';
-import { EdaProject } from '../interfaces/EdaProject';
+import { PageActions, PageHeader, PageLayout } from '../../../../framework';
+import { Scrollable } from '../../../../framework/components/Scrollable';
+import { TableDetails } from '../../../../framework/PageTable/PageTableDetails';
+import { useSettings } from '../../../../framework/Settings';
+import { useGet } from '../../../common/useItem';
+import { RouteE } from '../../../Routes';
+import { EdaProject } from '../../interfaces/EdaProject';
 import { useProjectActions } from './hooks/useProjectActions';
 import { useProjectColumns } from './hooks/useProjectColumns';
