+++ frontend/eda/Resources/projects/Projects.tsx
@@ -1,11 +1,11 @@
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { TablePage } from '../../../framework';
-import { useInMemoryView } from '../../../framework/useInMemoryView';
-import { useGet } from '../../common/useItem';
-import { idKeyFn } from '../../hub/usePulpView';
-import { RouteE } from '../../Routes';
-import { EdaProject } from '../interfaces/EdaProject';
+import { TablePage } from '../../../../framework';
+import { useInMemoryView } from '../../../../framework/useInMemoryView';
+import { useGet } from '../../../common/useItem';
+import { idKeyFn } from '../../../hub/usePulpView';
+import { RouteE } from '../../../Routes';
+import { EdaProject } from '../../interfaces/EdaProject';
 import { useProjectActions } from './hooks/useProjectActions';
 import { useProjectColumns } from './hooks/useProjectColumns';
 import { useProjectFilters } from './hooks/useProjectFilters';
